testing/web-platform/tests/html/cross-origin/null.tentative.html

<!doctype html>
<title>Cross-Origin header and nested navigable resource without such header</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
async_test(t => {
  const frame = document.createElement("iframe");
  frame.onload = t.step_func_done(() => {
    assert_not_equals(frame.contentDocument, null, "The frame should actually load");
  });
  frame.src = "/common/blank.html";
  document.body.append(frame);
  assert_equals(frame.contentDocument.body.localName, "body");
}, "Top-level with null policy: navigating a frame to a null policy should work.");

async_test(t => {
  const frame = document.createElement("iframe");
  let firstNavOk = false;
  frame.onload = t.step_func(() => {
    assert_not_equals(frame.contentDocument, null);
    firstNavOk = true;
  });
  t.step_timeout(() => {
    assert_equals(firstNavOk, true, "The initial load should work");
    assert_equals(frame.contentDocument, null, "Navigation to null policy should fail");
    t.done();
  }, 500);
  frame.src = "resources/navigate_anonymous.sub.html?to=/common/blank.html";
  document.body.append(frame);
  assert_equals(frame.contentDocument.body.localName, "body");
}, "Top-level with null policy: navigating a frame from use-credentials policy to a null policy should fail.");

async_test(t => {
  const frame = document.createElement("iframe");
  let firstNavOk = false;
  frame.onload = t.step_func(() => {
    assert_not_equals(frame.contentDocument, null);
    firstNavOk = true;
  });
  t.step_timeout(() => {
    assert_equals(firstNavOk, true, "The initial load should work");
    assert_equals(frame.contentDocument, null, "Navigation to null policy should fail");
    t.done();
  }, 500);
  frame.src = "resources/navigate_anonymous.sub.html?to=/common/blank.html";
  document.body.append(frame);
  assert_equals(frame.contentDocument.body.localName, "body");
}, "Top-level with null policy: navigating a frame from anonymous policy to a null policy should fail.");

async_test(t => {
  let w = window.open(`resources/navigate_null.sub.html?to=navigate_anonymous.sub.html`, "window_name");

  t.add_cleanup(() => w.close());

  t.step_timeout(() => {
    w.history.back();
    t.step_timeout(() => {
      assert_not_equals(w.document, null);
      t.done();
    }, 500);
  }, 500);
}, "Top-level with null policy: navigating a frame back from a blocked page should work.");


</script>
